+++ kafka-streams/kafka-streams-static-partitioning/src/main/java/org/apache/kafka/streams/processor/internals/OverridableStreamsPartitionAssignor.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.streams.processor.internals;
 
+// SUPPRESS CHECKSTYLE:OFF JavadocMethodRegex
 // SUPPRESS CHECKSTYLE:OFF ImportOrder
 // SUPPRESS CHECKSTYLE:OFF LineLength
 // SUPPRESS CHECKSTYLE:OFF ModifierOrder
@@ -25,23 +26,12 @@ package org.apache.kafka.streams.processor.internals;
 // SUPPRESS CHECKSTYLE:OFF NestedForDepth
 // SUPPRESS CHECKSTYLE:OFF JavadocStyle
 // SUPPRESS CHECKSTYLE:OFF NestedForDepth
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicBoolean;
+// SUPPRESS CHECKSTYLE:OFF UnusedImports
 
 import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.consumer.internals.PartitionAssignor;
+import org.apache.kafka.clients.consumer.internals.PartitionAssignor.Assignment;
+import org.apache.kafka.clients.consumer.internals.PartitionAssignor.Subscription;
 import org.apache.kafka.common.Cluster;
 import org.apache.kafka.common.Configurable;
 import org.apache.kafka.common.KafkaException;
@@ -56,8 +46,6 @@ import org.apache.kafka.streams.errors.StreamsException;
 import org.apache.kafka.streams.errors.TaskAssignmentException;
 import org.apache.kafka.streams.processor.PartitionGrouper;
 import org.apache.kafka.streams.processor.TaskId;
-import org.apache.kafka.clients.consumer.internals.PartitionAssignor.Assignment;
-import org.apache.kafka.clients.consumer.internals.PartitionAssignor.Subscription;
 import org.apache.kafka.streams.processor.internals.assignment.AssignmentInfo;
 import org.apache.kafka.streams.processor.internals.assignment.ClientState;
 import org.apache.kafka.streams.processor.internals.assignment.StickyTaskAssignor;
@@ -66,24 +54,65 @@ import org.apache.kafka.streams.processor.internals.assignment.TaskAssignor;
 import org.apache.kafka.streams.state.HostInfo;
 import org.slf4j.Logger;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
+import java.util.concurrent.atomic.AtomicInteger;
+
 import static org.apache.kafka.common.utils.Utils.getHost;
 import static org.apache.kafka.common.utils.Utils.getPort;
 
-//Note: The following class is copied from Kafka Streams 2.0.0 with the only changes being
+//Note: The following class is copied from Kafka Streams 2.2.0 with the only changes being
 //      those that allow overriding TaskAssignor (see 'TWITTER CHANGED' comment)
 public class OverridableStreamsPartitionAssignor implements PartitionAssignor, Configurable {
 
   private final static int UNKNOWN = -1;
-  public final static int NOT_AVAILABLE = -2;
   private final static int VERSION_ONE = 1;
   private final static int VERSION_TWO = 2;
   private final static int VERSION_THREE = 3;
+  private final static int VERSION_FOUR = 4;
   private final static int EARLIEST_PROBEABLE_VERSION = VERSION_THREE;
-  private int minReceivedMetadataVersion = UNKNOWN;
-  protected Set<Integer> supportedVersions = new HashSet<>();
+  protected final Set<Integer> supportedVersions = new HashSet<>();
 
   private Logger log;
   private String logPrefix;
+  public enum Error {
+    NONE(0),
+    INCOMPLETE_SOURCE_TOPIC_METADATA(1),
+    VERSION_PROBING(2);
+
+    private final int code;
+
+    Error(final int code) {
+      this.code = code;
+    }
+
+    public int code() {
+      return code;
+    }
+
+    public static Error fromCode(final int code) {
+      switch (code) {
+        case 0:
+          return NONE;
+        case 1:
+          return INCOMPLETE_SOURCE_TOPIC_METADATA;
+        case 2:
+          return VERSION_PROBING;
+        default:
+          throw new IllegalArgumentException("Unknown error code: " + code);
+      }
+    }
+  }
 
   private static class AssignedPartition implements Comparable<AssignedPartition> {
     public final TaskId taskId;
@@ -181,17 +210,19 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
     }
   }
 
-  protected static final Comparator<TopicPartition> PARTITION_COMPARATOR = new Comparator<TopicPartition>() {
-    @Override
-    public int compare(final TopicPartition p1,
-                       final TopicPartition p2) {
-      final int result = p1.topic().compareTo(p2.topic());
+  private static final class InternalStreamsConfig extends StreamsConfig {
+    private InternalStreamsConfig(final Map<?, ?> props) {
+      super(props, false);
+    }
+  }
 
-      if (result != 0) {
-        return result;
-      } else {
-        return Integer.compare(p1.partition(), p2.partition());
-      }
+  protected static final Comparator<TopicPartition> PARTITION_COMPARATOR = (p1, p2) -> {
+    final int result = p1.topic().compareTo(p2.topic());
+
+    if (result != 0) {
+      return result;
+    } else {
+      return Integer.compare(p1.partition(), p2.partition());
     }
   };
 
@@ -200,7 +231,7 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
 
   private TaskManager taskManager;
   private PartitionGrouper partitionGrouper;
-  private AtomicBoolean versionProbingFlag;
+  private AtomicInteger assignmentErrorCode;
 
   protected int usedSubscriptionMetadataVersion = SubscriptionInfo.LATEST_SUPPORTED_VERSION;
 
@@ -223,7 +254,7 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
    */
   @Override
   public void configure(final Map<String, ?> configs) {
-    final StreamsConfig streamsConfig = new StreamsConfig(configs);
+    final StreamsConfig streamsConfig = new InternalStreamsConfig(configs);
 
     // Setting the logger with the passed in client thread name
     logPrefix = String.format("stream-thread [%s] ", streamsConfig.getString(CommonClientConfigs.CLIENT_ID_CONFIG));
@@ -265,20 +296,20 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
 
     taskManager = (TaskManager) o;
 
-    final Object o2 = configs.get(StreamsConfig.InternalConfig.VERSION_PROBING_FLAG);
-    if (o2 == null) {
-      final KafkaException fatalException = new KafkaException("VersionProbingFlag is not specified");
+    final Object ai = configs.get(StreamsConfig.InternalConfig.ASSIGNMENT_ERROR_CODE);
+    if (ai == null) {
+      final KafkaException fatalException = new KafkaException("assignmentErrorCode is not specified");
       log.error(fatalException.getMessage(), fatalException);
       throw fatalException;
     }
 
-    if (!(o2 instanceof AtomicBoolean)) {
-      final KafkaException fatalException = new KafkaException(String.format("%s is not an instance of %s", o2.getClass().getName(), AtomicBoolean.class.getName()));
+    if (!(ai instanceof AtomicInteger)) {
+      final KafkaException fatalException = new KafkaException(String.format("%s is not an instance of %s",
+          ai.getClass().getName(), AtomicInteger.class.getName()));
       log.error(fatalException.getMessage(), fatalException);
       throw fatalException;
     }
-
-    versionProbingFlag = (AtomicBoolean) o2;
+    assignmentErrorCode = (AtomicInteger) ai;
 
     numStandbyReplicas = streamsConfig.getInt(StreamsConfig.NUM_STANDBY_REPLICAS_CONFIG);
 
@@ -335,6 +366,26 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
     return new Subscription(new ArrayList<>(topics), data.encode());
   }
 
+  Map<String, Assignment> errorAssignment(final Map<UUID, ClientMetadata> clientsMetadata,
+                                          final String topic,
+                                          final int errorCode) {
+    log.error("{} is unknown yet during rebalance," +
+        " please make sure they have been pre-created before starting the Streams application.", topic);
+    final Map<String, Assignment> assignment = new HashMap<>();
+    for (final ClientMetadata clientMetadata : clientsMetadata.values()) {
+      for (final String consumerId : clientMetadata.consumers) {
+        assignment.put(consumerId, new Assignment(
+            Collections.emptyList(),
+            new AssignmentInfo(AssignmentInfo.LATEST_SUPPORTED_VERSION,
+                Collections.emptyList(),
+                Collections.emptyMap(),
+                Collections.emptyMap(),
+                errorCode).encode()
+        ));
+      }
+    }
+    return assignment;
+  }
   /*
    * This assigns tasks to consumer clients in the following steps.
    *
@@ -362,7 +413,8 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
     final Map<UUID, ClientMetadata> clientsMetadata = new HashMap<>();
     final Set<String> futureConsumers = new HashSet<>();
 
-    minReceivedMetadataVersion = SubscriptionInfo.LATEST_SUPPORTED_VERSION;
+    int minReceivedMetadataVersion = SubscriptionInfo.LATEST_SUPPORTED_VERSION;
+
     supportedVersions.clear();
     int futureMetadataVersion = UNKNOWN;
     for (final Map.Entry<String, Subscription> entry : subscriptions.entrySet()) {
@@ -425,6 +477,12 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
 
     final Map<String, InternalTopicMetadata> repartitionTopicMetadata = new HashMap<>();
     for (final InternalTopologyBuilder.TopicsInfo topicsInfo : topicGroups.values()) {
+      for (final String topic : topicsInfo.sourceTopics) {
+        if (!topicsInfo.repartitionSourceTopics.keySet().contains(topic) &&
+            !metadata.topics().contains(topic)) {
+          return errorAssignment(clientsMetadata, topic, Error.INCOMPLETE_SOURCE_TOPIC_METADATA.code);
+        }
+      }
       for (final InternalTopicConfig topic: topicsInfo.repartitionSourceTopics.values()) {
         repartitionTopicMetadata.put(topic.name(), new InternalTopicMetadata(topic));
       }
@@ -454,12 +512,9 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
                     numPartitionsCandidate = repartitionTopicMetadata.get(sourceTopicName).numPartitions;
                   } else {
                     numPartitionsCandidate = metadata.partitionCountForTopic(sourceTopicName);
-                    if (numPartitionsCandidate == null) {
-                      repartitionTopicMetadata.get(topicName).numPartitions = NOT_AVAILABLE;
-                    }
                   }
 
-                  if (numPartitionsCandidate != null && numPartitionsCandidate > numPartitions) {
+                  if (numPartitionsCandidate > numPartitions) {
                     numPartitions = numPartitionsCandidate;
                   }
                 }
@@ -600,7 +655,7 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
 
     // construct the global partition assignment per host map
     final Map<HostInfo, Set<TopicPartition>> partitionsByHostState = new HashMap<>();
-    if (minReceivedMetadataVersion == 2 || minReceivedMetadataVersion == 3) {
+    if (minReceivedMetadataVersion >= 2) {
       for (final Map.Entry<UUID, ClientMetadata> entry : clientsMetadata.entrySet()) {
         final HostInfo hostInfo = entry.getValue().hostInfo;
 
@@ -681,7 +736,7 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
         // finally, encode the assignment before sending back to coordinator
         assignment.put(consumer, new Assignment(
             activePartitions,
-            new AssignmentInfo(minUserMetadataVersion, active, standby, partitionsByHostState).encode()));
+            new AssignmentInfo(minUserMetadataVersion, active, standby, partitionsByHostState, 0).encode()));
       }
     }
 
@@ -721,7 +776,8 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
                 minUserMetadataVersion,
                 activeTasks,
                 standbyTasks,
-                partitionsByHostState)
+                partitionsByHostState,
+                0)
                 .encode()
         ));
       }
@@ -767,6 +823,11 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
     Collections.sort(partitions, PARTITION_COMPARATOR);
 
     final AssignmentInfo info = AssignmentInfo.decode(assignment.userData());
+    if (info.errCode() != Error.NONE.code) {
+      // set flag to shutdown streams app
+      assignmentErrorCode.set(info.errCode());
+      return;
+    }
     final int receivedAssignmentMetadataVersion = info.version();
     final int leaderSupportedVersion = info.latestSupportedVersion();
 
@@ -793,7 +854,7 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
         usedSubscriptionMetadataVersion = leaderSupportedVersion;
       }
 
-      versionProbingFlag.set(true);
+      assignmentErrorCode.set(Error.VERSION_PROBING.code);
       return;
     }
 
@@ -822,6 +883,18 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
           usedSubscriptionMetadataVersion = leaderSupportedVersion;
         }
         processVersionThreeAssignment(info, partitions, activeTasks, topicToPartitionInfo);
+        partitionsByHost = info.partitionsByHost();
+        break;
+      case VERSION_FOUR:
+        if (leaderSupportedVersion > usedSubscriptionMetadataVersion) {
+          log.info("Sent a version {} subscription and group leader's latest supported version is {}. " +
+                  "Upgrading subscription metadata version to {} for next rebalance.",
+              usedSubscriptionMetadataVersion,
+              leaderSupportedVersion,
+              leaderSupportedVersion);
+          usedSubscriptionMetadataVersion = leaderSupportedVersion;
+        }
+        processVersionFourAssignment(info, partitions, activeTasks, topicToPartitionInfo);
         partitionsByHost = info.partitionsByHost();
         break;
       default:
@@ -877,6 +950,13 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
     processVersionTwoAssignment(info, partitions, activeTasks, topicToPartitionInfo);
   }
 
+  private void processVersionFourAssignment(final AssignmentInfo info,
+                                            final List<TopicPartition> partitions,
+                                            final Map<TaskId, Set<TopicPartition>> activeTasks,
+                                            final Map<TopicPartition, PartitionInfo> topicToPartitionInfo) {
+    processVersionThreeAssignment(info, partitions, activeTasks, topicToPartitionInfo);
+  }
+
   // for testing
   protected void processLatestVersionAssignment(final AssignmentInfo info,
                                                 final List<TopicPartition> partitions,
@@ -900,9 +980,6 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
       final InternalTopicConfig topic = metadata.config;
       final int numPartitions = metadata.numPartitions;
 
-      if (numPartitions == NOT_AVAILABLE) {
-        continue;
-      }
       if (numPartitions < 0) {
         throw new StreamsException(String.format("%sTopic [%s] number of partitions not defined", logPrefix, topic.name()));
       }
@@ -928,9 +1005,12 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
 
   static class CopartitionedTopicsValidator {
     private final String logPrefix;
+    private final Logger log;
 
     CopartitionedTopicsValidator(final String logPrefix) {
       this.logPrefix = logPrefix;
+      final LogContext logContext = new LogContext(logPrefix);
+      log = logContext.logger(getClass());
     }
 
     void validate(final Set<String> copartitionGroup,
@@ -941,9 +1021,10 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
       for (final String topic : copartitionGroup) {
         if (!allRepartitionTopicsNumPartitions.containsKey(topic)) {
           final Integer partitions = metadata.partitionCountForTopic(topic);
-
           if (partitions == null) {
-            throw new org.apache.kafka.streams.errors.TopologyException(String.format("%sTopic not found: %s", logPrefix, topic));
+            final String str = String.format("%sTopic not found: %s", logPrefix, topic);
+            log.error(str);
+            throw new IllegalStateException(str);
           }
 
           if (numPartitions == UNKNOWN) {
@@ -953,9 +1034,6 @@ public class OverridableStreamsPartitionAssignor implements PartitionAssignor, C
             Arrays.sort(topics);
             throw new org.apache.kafka.streams.errors.TopologyException(String.format("%sTopics not co-partitioned: [%s]", logPrefix, Utils.join(Arrays.asList(topics), ",")));
           }
-        } else if (allRepartitionTopicsNumPartitions.get(topic).numPartitions == NOT_AVAILABLE) {
-          numPartitions = NOT_AVAILABLE;
-          break;
         }
       }