+++ kafka-streams/kafka-streams/src/main/scala/com/twitter/finatra/kafkastreams/internal/utils/TopologyReflectionUtils.scala
@@ -11,8 +11,7 @@ private[kafkastreams] object TopologyReflectionUtils {
   def isStateless(topology: Topology): Boolean = {
     val internalTopologyBuilder = getInternalTopologyBuilder(topology)
 
-    internalTopologyBuilder.allStateStoreName().isEmpty &&
-    internalTopologyBuilder.globalStateStores().isEmpty
+    internalTopologyBuilder.getStateStores.isEmpty
   }
 
   private def getInternalTopologyBuilder(topology: Topology): InternalTopologyBuilder = {