+++ kafka-streams/kafka-streams/src/main/scala/org/apache/kafka/streams/state/internals/WindowKeySchemaExtractor.scala
@@ -1,0 +1,11 @@
+package org.apache.kafka.streams.state.internals
+
+object WindowKeySchemaExtractor {
+  def extractStoreTimestamp(binaryKey: Array[Byte]): Long = {
+    WindowKeySchema.extractStoreTimestamp(binaryKey)
+  }
+
+  def extractStoreKeyBytes(binaryKey: Array[Byte]): Array[Byte] = {
+    WindowKeySchema.extractStoreKeyBytes(binaryKey)
+  }
+}