+++ kafka-streams/kafka-streams/src/test/scala/com/twitter/finatra/kafkastreams/integration/stateless/VerifyFailureServer.scala
@@ -4,7 +4,7 @@ import com.twitter.finatra.kafka.serde.ScalaSerdes
 import com.twitter.finatra.kafkastreams.utils.StatelessKafkaStreamsTwitterServer
 import org.apache.kafka.common.serialization.Serdes
 import org.apache.kafka.streams.StreamsBuilder
-import org.apache.kafka.streams.kstream.{Consumed, Materialized, Produced, Serialized}
+import org.apache.kafka.streams.kstream.{Consumed, Grouped, Materialized, Produced}
 
 class VerifyFailureServer extends StatelessKafkaStreamsTwitterServer {
 
@@ -13,7 +13,7 @@ class VerifyFailureServer extends StatelessKafkaStreamsTwitterServer {
     builder.asScala
       .stream("TextLinesTopic")(Consumed.`with`(Serdes.Bytes, Serdes.String))
       .flatMapValues(_.split(' '))
-      .groupBy((_, word) => word)(Serialized.`with`(Serdes.String, Serdes.String))
+      .groupBy((_, word) => word)(Grouped.`with`(Serdes.String, Serdes.String))
       .count()(Materialized.as("CountsStore"))
       .toStream
       .to("WordsWithCountsTopic")(Produced.`with`(Serdes.String, ScalaSerdes.Long))