+++ CHANGELOG.rst
@@ -22,6 +22,13 @@ Added
 Changed
 ~~~~~~~
 
+* finatra-kafka: Upgraded kafka libraries from 2.0.0 to 2.2.0.
+  - `Kafka 2.0.1 Release Notes <https://archive.apache.org/dist/kafka/2.0.1/RELEASE_NOTES.html>`__
+  - `Kafka 2.1.0 Release Notes <https://archive.apache.org/dist/kafka/2.1.0/RELEASE_NOTES.html>`__
+  - `Kafka 2.1.1 Release Notes <https://archive.apache.org/dist/kafka/2.1.1/RELEASE_NOTES.html>`__
+  - `Kafka 2.2.0 Release Notes <https://archive.apache.org/dist/kafka/2.2.0/RELEASE_NOTES.html>`__
+  ``PHAB_ID=D248171``
+
 * finatra-thrift: Removed `c.t.finatra.thrift.exceptions.FinatraThriftExceptionMapper`,
   `c.t.finatra.thrift.filters.ClientIdAcceptlistFilter`,
   `c.t.finatra.thrift.modules.ClientIdAcceptlistModule`,